Some initial work on DAGKnight. As in the part where the computation of DAGKnight K is done.
The calculation where we select the largest 𝑘-cluster for each value of 𝑘, and pick the minimal 𝑘 whose maximizing cluster covers 50% of the DAG. We thus utilize the honest majority assumption to recognize a subset of blocks that are sufficient to counter an attack. In this way, we avoid the need to know 𝑘 in advance, and allow the protocol to self-adjust to the real time latency.

This still needs to be throughly tested and implemented accordingly to replace GhostDAG.